refactor: Extract useGeographyCapitalsGame and useGeographyContinentsGame from gameHooksMap.ts
- Create hooks/games/useGeographyCapitalsGame.ts — reads store items, builds
  pronunciation map (speaks country name not capital), calls useBaseGame
- Create hooks/games/useGeographyContinentsGame.ts — same pattern but adds
  uniqueByField: 'color' so options always come from different continents
- Update gameHooksMap.ts — remove the two inline hook lambdas; import the
  extracted named hooks instead (same pattern as useNatureSoundsGame)
  Keeps useGameTypeStore import for the generic g() helper
